feat: add changeView.ts with class for change size products
